Library System/Back-end/LibrarySystem/requirements.txt
Django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
mysqlclient==2.2.0
PyJWT==2.8.0 